CODEBASE/phase2_analysis/requirements.txt
typer==0.12.5
pydantic==2.9.2
orjson==3.10.7
transformers==4.46.0
torch==2.4.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
ollama==0.2.1
rich==13.9.3
pyyaml==6.0.2

